requirements.txt
asyncpg==0.30.0
fastapi==0.116.1
groq==0.31.0
python-dotenv==1.1.1
twilio==9.6.5
uvicorn==0.35.0
websockets==15.0.1
jinja2==3.1.6
asyncio==4.0.0